split indexers out into separate file
