+++ src/pyterrier_pisa/__init__.py
@@ -1,14 +1,10 @@
 import numpy as np
-from . import _pisathon
 import json
-import shutil
 import sys
-import pandas as pd
 from pathlib import Path
 import tempfile
 import os
 import more_itertools
-import threading
 from warnings import warn
 from typing import Optional, Union, List
 from enum import Enum
@@ -16,8 +12,9 @@ from collections import Counter
 import pyterrier as pt
 from pyterrier.datasets import Dataset
 import functools
-from collections import defaultdict
 import ir_datasets
+from . import _pisathon
+from .indexers import PisaIndexer, PisaToksIndexer, PisaIndexingMode
 
 _logger = ir_datasets.log.easy()
 
@@ -79,12 +76,6 @@ class PisaStopwords(Enum):
   """
   terrier = 'terrier'
   none = 'none'
-
-
-class PisaIndexingMode(Enum):
-  create = 'create'
-  overwrite = 'overwrite'
-  # append?
 
 
 PISA_INDEX_DEFAULTS = {
@@ -153,6 +144,7 @@ class PisaIndex(pt.Indexer):
         warn(f'text_field not specified; indexing all str fields: {text_field}')
 
     mode = PisaIndexingMode.overwrite if self.overwrite else PisaIndexingMode.create
+    warn(f'{mode} {type(mode)}')
 
     if isinstance(text_field, str) and isinstance(first_doc[text_field], dict):
       return self.toks_indexer(text_field, mode=mode).index(it)
@@ -251,112 +243,6 @@ class PisaIndex(pt.Indexer):
     if PisaStemmer(self.stemmer) != PisaStemmer.none:
       raise ValueError("To index from dicts, you must set stemmer='none'")
     return PisaToksIndexer(self.path, text_field, mode, threads=threads or self.threads, batch_size=self.batch_size)
-
-
-class PisaIndexer(pt.Indexer):
-  def __init__(self, path, text_field, mode=PisaIndexingMode.create, stemmer=PISA_INDEX_DEFAULTS['stemmer'], threads=1, batch_size=100_000):
-    self.path = path
-    self.text_field = text_field
-    self.mode = PisaIndexingMode(mode)
-    self.stemmer = PisaStemmer(stemmer)
-    self.threads = threads
-    self.batch_size = batch_size
-
-  def index(self, it):
-    path = Path(self.path)
-    if PisaIndex.built(self):
-      if PisaIndexingMode(self.mode) == PisaIndexingMode.overwrite:
-        warn(f'Removing {str(path)}')
-        shutil.rmtree(path)
-      else:
-        raise RuntimeError('A PISA index already exists at {self.path}. If you want to overwrite it, set mode="overwrite"')
-    if not path.exists():
-      path.mkdir(parents=True, exist_ok=True)
-
-    self._index(it)
-
-    with open(path/'pt_pisa_config.json', 'wt') as fout:
-      json.dump({
-        'stemmer': self.stemmer.value,
-      }, fout)
-    return PisaIndex(self.path, batch_size=self.batch_size, stemmer=self.stemmer, text_field=self.text_field, threads=self.threads)
-
-  def _index(self, it):
-    with tempfile.TemporaryDirectory() as d:
-      fifo = os.path.join(d, 'fifo')
-      os.mkfifo(fifo)
-      threading.Thread(target=self._write_fifo, args=(it, fifo, self.text_field), daemon=True).start()
-      _pisathon.index(fifo, str(self.path), '' if self.stemmer == PisaStemmer.none else self.stemmer.value, self.batch_size, self.threads)
-
-  def _write_fifo(self, it, fifo, text_field):
-    with open(fifo, 'wt') as fout:
-      if isinstance(text_field, str):
-        text_field = [text_field]
-      for doc in it:
-        docno, text = doc['docno'], ' '.join(doc[f] for f in text_field)
-        text = text.replace('\n', ' ').replace('\r', ' ') # any other cleanup?
-        fout.write(f'{docno} {text}\n')
-
-
-class PisaToksIndexer(PisaIndexer):
-  def __init__(self, path, text_field, mode=PisaIndexingMode.create, threads=1, batch_size=100_000):
-    super().__init__(path, text_field, mode, PisaStemmer.none, threads, batch_size=batch_size)
-
-  def _index(self, it):
-    lexicon = {}
-    docid = 0
-    path = Path(self.path)
-    with (path/'fwd.documents').open('wt') as f_docs, (path/'fwd.terms').open('wt') as f_lex:
-      for bidx, batch in enumerate(more_itertools.chunked(it, self.batch_size)):
-        _logger.info(f'inverting batch {bidx}: documents [{docid},{docid+len(batch)})')
-        inv_did = defaultdict(list)
-        inv_score = defaultdict(list)
-        lens = []
-        for doc in batch:
-          l = 0
-          f_docs.write(doc['docno']+'\n')
-          for term, score in doc[self.text_field].items():
-            score = int(score)
-            if score <= 0:
-              continue
-            l += score
-            if term not in lexicon:
-              lexicon[term] = len(lexicon)
-              f_lex.write(term+'\n')
-            inv_did[lexicon[term]].append(docid)
-            inv_score[lexicon[term]].append(int(score))
-          lens.append(l)
-          docid += 1
-        with (path/f'inv.batch.{bidx}.docs').open('wb') as f_did, (path/f'inv.batch.{bidx}.freqs').open('wb') as f_score, (path/f'inv.batch.{bidx}.sizes').open('wb') as f_len:
-          f_did.write(np.array([1, len(batch)], dtype=np.uint32).tobytes())
-          for i in range(len(lexicon)):
-            if i in inv_did:
-              l = len(inv_did[i])
-              f_did.write(np.array([l] + inv_did[i], dtype=np.uint32).tobytes())
-              f_score.write(np.array([l] + inv_score[i], dtype=np.uint32).tobytes())
-          f_len.write(np.array([len(lens)] + lens, dtype=np.uint32).tobytes())
-    _pisathon.merge_inv(str(path/'inv'), bidx+1, len(lexicon))
-    (path/'inv.docs').rename(path/'inv.docs.tmp')
-    (path/'inv.freqs').rename(path/'inv.freqs.tmp')
-    in_docs = np.memmap(path/'inv.docs.tmp', mode='r', dtype=np.uint32)
-    in_freqs = np.memmap(path/'inv.freqs.tmp', mode='r', dtype=np.uint32)
-    with (path/'fwd.terms').open('wt') as f_lex, (path/'inv.docs').open('wb') as f_docs, (path/'inv.freqs').open('wb') as f_freqs:
-      f_docs.write(in_docs[:2].tobytes())
-      in_docs = in_docs[2:]
-      offsets_lens = []
-      i = 0
-      while i < in_docs.shape[0]:
-        offsets_lens.append((i, in_docs[i]+1))
-        i += in_docs[i] + 1
-      print(len(offsets_lens))
-      for term in _logger.pbar(sorted(lexicon), desc='re-mapping term ids'):
-        f_lex.write(f'{term}\n')
-        i = lexicon[term]
-        start, l = offsets_lens[i]
-        f_docs.write(in_docs[start:start+l])
-        f_freqs.write(in_freqs[start:start+l])
-    _pisathon.build_binlex(str(path/'fwd.documents'), str(path/'fwd.doclex'))
-    _pisathon.build_binlex(str(path/'fwd.terms'), str(path/'fwd.termlex'))
 
 
 class PisaRetrieve(pt.Transformer):
